fix(move): route CLAMM fee swaps through StdAddr
