+++ x/move/keeper/dex.go
@@ -292,7 +292,35 @@ func (k DexKeeper) SwapToBase(
 		if ok, err := clammKeeper.HasPool(ctx, metadataLP); err != nil {
 			return err
 		} else if ok {
-			return clammKeeper.SwapToBase(ctx, vmAddr, metadataLP, metadataQuote, quoteCoin.Amount)
+			baseDenom := params.BaseDenom
+			prevBaseBalance, err := k.moveBankKeeper.GetBalance(ctx, types.StdAddr, baseDenom)
+			if err != nil {
+				return err
+			}
+
+			// CLAMM cannot swap directly from blocked recipients such as the fee collector,
+			// so route the quote through StdAddr, execute the swap there, then send only
+			// the base-denom delta back to the original caller.
+			if err := k.moveBankKeeper.SendCoin(ctx, addr, types.StdAddr, quoteCoin.Denom, quoteCoin.Amount); err != nil {
+				return err
+			}
+			if err := clammKeeper.SwapToBase(ctx, vmtypes.StdAddress, metadataLP, metadataQuote, quoteCoin.Amount); err != nil {
+				return err
+			}
+
+			postBaseBalance, err := k.moveBankKeeper.GetBalance(ctx, types.StdAddr, baseDenom)
+			if err != nil {
+				return err
+			}
+
+			if postBaseBalance.GT(prevBaseBalance) {
+				baseBalanceDiff := postBaseBalance.Sub(prevBaseBalance)
+				if err := k.moveBankKeeper.SendCoin(ctx, types.StdAddr, addr, params.BaseDenom, baseBalanceDiff); err != nil {
+					return err
+				}
+			}
+
+			return nil
 		}
 	}
